Fix: small line spacing sometimes causing text to pile up in the last line.
-[NSString sizeWithFont:constrainedToSize:...] and
-[NSString FXLabel_linesWithFont:...] methods can disagree about the space
required to draw the label, because the former roundf()s the calculated line
spacing while the latter doesn't. If they disagree, FXLabel_linesWithFont
appends the excess text to the last line it calculates will fit.

Both methods now roundf() the calculated extra line spacing, which seems
to resolve the issue.